Extract MigrationIsInProgress to a module
